refactor: 메소드명 및 디렉토리 구조 개선
코드 가독성을 높이기 위해 메소드를 추가하고 클래스를 디렉토리에 분류했습니다.
