+++ src/main/java/calculator/input/InputHandler.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.input;
 
 import camp.nextstep.edu.missionutils.Console;
