+++ src/main/java/calculator/input/InputValidator.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.input;
 
 import java.util.ArrayList;
 import java.util.List;
