+++ src/main/java/calculator/service/DelimiterProcessor.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.service;
 
 public class DelimiterProcessor {
