+++ src/main/java/calculator/service/ResultPrinter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.service;
 
 public class ResultPrinter {
